Login: Move the login component out of the auth section
