+++ assets/stylesheets/_components.scss
@@ -17,7 +17,7 @@
 @import 'account-recovery/reset-password/transaction-id-form/style';
 @import 'account-recovery/style';
 @import 'auth/style';
-@import 'auth/wp-login/style';
+@import 'login/wp-login/style';
 @import 'blocks/app-promo/style';
 @import 'blocks/author-compact-profile/style';
 @import 'blocks/author-selector/style';
