+++ client/login/wp-login/index.js
